% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alphacodes}
\alias{alphacodes}
\title{Alpha codes (4-letter and 6-letter) for North American bird species.}
\format{
A data frame with 2266 rows and 5 variables:
\describe{
\item{source}{Name of the source of the alpha codes. Institute for Bird Populations (IBP). Bird Banding Lab (BBL)}
\item{alpha4}{4-letter alpha code associated with the \code{source}}
\item{alpha6}{6-letter alpha code associated with the \code{source}}
\item{commonname}{English common name provided by the \code{source}}
\item{sciname}{Latin scientific name provided by the \code{source}}
}
}
\source{
IBP codes must currently be manually downloaded from\url{https://www.birdpop.org/pages/birdSpeciesCodes.php}

BBL codes are not currently published, but will be included in the same data object as the IBP (here)
}
\usage{
alphacodes
}
\description{
A dataset containing alpha codes from (currently only) the IBP. (Future addition BBL)
}
\keyword{datasets}
